+++ src/solver.cpp
@@ -357,7 +357,7 @@ retStateT Solver::resolveConflict() {
 bool Solver::bcp() {
 // the asserted literal has been set, so we start
 // bcp on that literal
-	unsigned start_ofs = literal_stack_.size() - 1;
+	size_t start_ofs = literal_stack_.size() - 1;
 
 //BEGIN process unit clauses
 	for (auto lit : unit_clauses_)
@@ -372,8 +372,8 @@ bool Solver::bcp() {
 	return bSucceeded;
 }
 
-bool Solver::BCP(unsigned start_at_stack_ofs) {
-	for (unsigned int i = start_at_stack_ofs; i < literal_stack_.size(); i++) {
+bool Solver::BCP(size_t start_at_stack_ofs) {
+	for (size_t i = start_at_stack_ofs; i < literal_stack_.size(); i++) {
 		LiteralID unLit = literal_stack_[i].neg();
 		//BEGIN Propagate Bin Clauses
 		for (auto bt = literal(unLit).binary_links_.begin();
